requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
python-dotenv
Pillow
psycopg2-binary
drf-spectacular
#django-filter
